rework hashdata into more separated namespaces
