+++ libs/sr/dm-localization/src/executors/gen-locale/lib/charHashData/character.ts
@@ -1,0 +1,32 @@
+export const charHashData = {
+  label: {
+    one: '-1435714104', // Character
+    many: '-676993900', // Characters
+  },
+  eidolonLvl: '1102721995', // "Eidolon Level:"
+  traces: '1252031000',
+  type: {
+    physical: '-1101149179',
+    quantum: '88513016',
+    lightning: '-250814279',
+    ice: '-756819502',
+    wind: '-1513894875',
+    fire: '-1474263915',
+    imaginary: '118034640',
+  },
+  ability: {
+    basic: '216145179',
+    skill: '216145178',
+    ult: '216145177',
+    talent: '216145176',
+    technique: '216145173',
+  },
+  dmgType: {
+    single: '-811923716',
+    aoe: '-408639189',
+    bounce: '-1571438603',
+    blast: '1157444752',
+    support: '1560729279',
+    impair: '397929865',
+  },
+}
