+++ libs/sr/dm-localization/src/executors/gen-locale/lib/charHashData/index.ts
@@ -1,0 +1,2 @@
+export * from './allCharacters'
+export * from './character'
