--- libs/sr/dm-localization/src/executors/gen-locale/lib/commonHashData.ts
@@ -1,106 +1,0 @@
-const charHashData = {
-  eidolonLvl: '1102721995', // "Eidolon Level:"
-  traces: '1252031000',
-  type: {
-    physical: '-1101149179',
-    quantum: '88513016',
-    lightning: '-250814279',
-    ice: '-756819502',
-    wind: '-1513894875',
-    fire: '-1474263915',
-    imaginary: '118034640',
-  },
-  ability: {
-    basic: '216145179',
-    skill: '216145178',
-    ult: '216145177',
-    talent: '216145176',
-    technique: '216145173',
-  },
-  dmgType: {
-    single: '-811923716',
-    aoe: '-408639189',
-    bounce: '-1571438603',
-    blast: '1157444752',
-    support: '1560729279',
-    impair: '397929865',
-  },
-}
-
-const relicHashData = {
-  set: '1604210092',
-  mainStat: '-338586359',
-  slot: '-1409515362',
-  slotKey: {
-    head: '1826268247',
-    hands: '-2144127003',
-    body: '1799740313',
-    feet: '1154536283',
-    sphere: '1873322474',
-    rope: '74129613',
-  },
-}
-
-const lightConeHashData = {
-  superimposition: '-2008197403',
-}
-
-const teamHashData = {
-  name: '858715661', // Team Name
-  editTeam: '-756448810',
-}
-
-export const commonHashData = {
-  char: charHashData,
-  relic: relicHashData,
-  lc: lightConeHashData,
-  team: teamHashData,
-
-  character: '-1435714104',
-  characters: '-676993900',
-  lightCone: '-514739711',
-  lightCones: '582917607',
-  relics: '-1823495414',
-  teamLabel: '-839489437',
-  level: '-636045037', // Level
-  lv: '982642653', // Lv.
-  rarity: '1898197464',
-  path: '-1952347303',
-  paths: '-815380608',
-  pathKey: {
-    Erudition: '-953720572',
-    Preservation: '-1357005099',
-    Abundance: '1371878256',
-    Nihility: '-1760289626',
-    Destruction: '209078842',
-    Harmony: '-194205685',
-    TheHunt: '612363369',
-  },
-  statKey: {
-    spd: '461357812',
-    spd_: '461357812',
-    hp: '-1544419067',
-    hp_: '-1544419067',
-    atk: '-1126979077',
-    atk_: '-1126979077',
-    def: '-615378225',
-    def_: '-615378225',
-    crit_: '528301800',
-    crit_dmg_: '-220753499',
-    brEffect_: '-1668395853',
-    eff_: '284076746',
-    eff_res_: '-952827954',
-    enerRegen_: '-1155541365',
-    heal_: '799727458',
-    physical_dmg_: '-950888303',
-    fire_dmg_: '1202183228',
-    ice_dmg_: '1448585423',
-    wind_dmg_: '2017136070',
-    lightning_dmg_: '235195444',
-    quantum_dmg_: '1819458027',
-    imaginary_dmg_: '1031064451',
-    dmg_: '-907763317',
-    weakness_: '-1297647890',
-    resPen_: '-538204877',
-  },
-}
