+++ libs/sr/dm-localization/src/executors/gen-locale/lib/hashData.ts
@@ -4,39 +4,96 @@ import {
   allLightConeKeys,
   allRelicSetKeys,
 } from '@genshin-optimizer/sr/consts'
-import { charHashData } from './charHashData'
-import { commonHashData } from './commonHashData'
-import { lightConeHashData } from './lightConeHashData'
-import { relicHashData } from './relicHashData'
+import { allCharHashData, charHashData } from './charHashData'
+import { allLightConeHashData, lightConeHashData } from './lightConeHashData'
+import { allRelicHashData, relicHashData } from './relicHashData'
 import { uiHashData } from './uiHashData'
 
 export type LanguageData = typeof HashData
 
 const charNames = Object.fromEntries(
-  Object.entries(charHashData).map(([key, data]) => [key, data.name])
+  Object.entries(allCharHashData).map(([key, data]) => [key, data.name])
 )
 verifyObjKeys(charNames, allCharacterDataKeys)
 
 const relicNames = Object.fromEntries(
-  Object.entries(relicHashData).map(([key, data]) => [key, data.setName])
+  Object.entries(allRelicHashData).map(([key, data]) => [key, data.setName])
 )
 verifyObjKeys(relicNames, allRelicSetKeys)
 
 const lightConeNames = Object.fromEntries(
-  Object.entries(lightConeHashData).map(([key, data]) => [key, data.name])
+  Object.entries(allLightConeHashData).map(([key, data]) => [key, data.name])
 )
 verifyObjKeys(lightConeNames, allLightConeKeys)
 
 export const HashData = {
-  char: charHashData,
+  char: allCharHashData,
   charNames,
+  characters: charHashData,
 
-  relic: relicHashData,
+  relic: allRelicHashData,
   relicNames,
+  relics: relicHashData,
 
-  lightCone: lightConeHashData,
+  lightCone: allLightConeHashData,
   lightConeNames,
+  lightCones: lightConeHashData,
 
-  common: commonHashData,
   ui: uiHashData,
+
+  common: {
+    level: '-636045037', // Level
+    lv: '982642653', // Lv.
+    rarity: '1898197464',
+  },
+
+  teams: {
+    label: '-839489437', // Team
+    name: '858715661', // Team Name
+    editTeam: '-756448810',
+  },
+
+  paths: {
+    label: {
+      one: '-1952347303', // Path
+      many: '-815380608', // Paths
+    },
+    pathKey: {
+      Erudition: '-953720572',
+      Preservation: '-1357005099',
+      Abundance: '1371878256',
+      Nihility: '-1760289626',
+      Destruction: '209078842',
+      Harmony: '-194205685',
+      TheHunt: '612363369',
+    },
+  },
+
+  statKey: {
+    spd: '461357812',
+    spd_: '461357812',
+    hp: '-1544419067',
+    hp_: '-1544419067',
+    atk: '-1126979077',
+    atk_: '-1126979077',
+    def: '-615378225',
+    def_: '-615378225',
+    crit_: '528301800',
+    crit_dmg_: '-220753499',
+    brEffect_: '-1668395853',
+    eff_: '284076746',
+    eff_res_: '-952827954',
+    enerRegen_: '-1155541365',
+    heal_: '799727458',
+    physical_dmg_: '-950888303',
+    fire_dmg_: '1202183228',
+    ice_dmg_: '1448585423',
+    wind_dmg_: '2017136070',
+    lightning_dmg_: '235195444',
+    quantum_dmg_: '1819458027',
+    imaginary_dmg_: '1031064451',
+    dmg_: '-907763317',
+    weakness_: '-1297647890',
+    resPen_: '-538204877',
+  },
 }
