+++ libs/sr/dm-localization/src/executors/gen-locale/lib/lightConeHashData/lightCone.ts
@@ -1,0 +1,7 @@
+export const lightConeHashData = {
+  label: {
+    one: '-514739711', // Light Cone
+    many: '582917607', // Light Cones
+  },
+  superimposition: '-2008197403',
+}
