+++ libs/sr/dm-localization/src/executors/gen-locale/lib/relicHashData/index.ts
@@ -1,0 +1,2 @@
+export * from './allRelics'
+export * from './relic'
