+++ libs/sr/dm-localization/src/executors/gen-locale/lib/relicHashData/relic.ts
@@ -1,0 +1,14 @@
+export const relicHashData = {
+  label: '-1823495414', // Relics
+  set: '1604210092',
+  mainStat: '-338586359',
+  slot: '-1409515362',
+  slotKey: {
+    head: '1826268247',
+    hands: '-2144127003',
+    body: '1799740313',
+    feet: '1154536283',
+    sphere: '1873322474',
+    rope: '74129613',
+  },
+}
